pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "glorb"
requires-python = ">=3.11"
license = "MIT"
keywords = []
authors = [
    { name = "Aarni Koskela", email = "akx@iki.fi" },
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
    "icosphere==0.1.3",
    "netCDF4==1.6.3",
    "numpy-stl==3.0.1",
    "diskcache==5.4.0",
    "tqdm",
]
dynamic = ["version"]

[project.urls]
Documentation = "https://github.com/akx/glorb#readme"
Issues = "https://github.com/akx/glorb/issues"
Source = "https://github.com/akx/glorb"

[tool.hatch.version]
path = "glorb/__init__.py"

[tool.hatch.envs.default]
dependencies = [
    "pytest",
    "pytest-cov",
]
[tool.hatch.envs.default.scripts]
cov = "pytest --cov-report=term-missing --cov-config=pyproject.toml --cov=glorb --cov=tests {args}"
no-cov = "cov --no-cov {args}"

[[tool.hatch.envs.test.matrix]]
python = ["311"]

[tool.coverage.run]
branch = true
parallel = true

[tool.coverage.report]
exclude_lines = [
    "no cov",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

[tool.ruff]
extend-select = ["COM"]
